requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
msal==1.29.0
Flask-Session==0.8.0
